+++ UI8/src/app/mud/file-info.ts
@@ -1,13 +1,31 @@
 export class FileInfo {
     lasturl : string = '';
-    filepath : string = '';
+    file : string = '';
+    path : string= '';
     filename : string = '';
     filetype : string = '';
-    content : string = '';
+    edditortype? : string = '';
+
+    newfile : boolean = true;
+    writeacl : boolean = false;
+    temporary : boolean = false;
     saveActive : boolean = false;
-    alreadyLoaded : boolean = false;
-    windowsId : string = undefined;
+    filesize : number = -1;
+    title: string ='';
+
+    content? : string = '';
+    oldContent? : string = '';
+
+    alreadyLoaded? : boolean = false;
+    windowsId? : string = undefined;
+    save01_start? = function(filepath,cb) {};
+    save02_url? = function(url) {};
+    save03_saved? = function(filepath,cb){};
+    save04_closing? = function(windowsid){};
+    save05_error? = function(windowsid,error){};
+    save06_success? = function(windowsid){};
     relateWindow? = function(wid){};    
     load? = function(cb) {};
-    save? = function(txt,cb,) {};
+    cancel01_start? = function(filepath,cb) {};
+    cancel02_end? = function(filepath) {};
 }