+++ UI8/src/app/mud/files.service.ts
@@ -1,6 +1,7 @@
 import { Injectable } from '@angular/core';
 import { FileInfo } from './file-info';
 import { HttpClient, HttpHeaders } from '@angular/common/http';
+import {NGXLogger} from 'ngx-logger';
 
 @Injectable({
   providedIn: 'root'
@@ -13,74 +14,67 @@ export class FilesService {
     
   }
 
-  processURL( url : string) : FileInfo {
-    const urlparts = url.split("=");
-    if (urlparts.length != 2) {
-      console.log("url mismatch:",url);
-      return;
-    }
-    const token = urlparts[1].split(".");
-    if (token.length!=3) {
-      console.log("token mismatch:",url);
-      return;
-    }
-    var fileinfo  :FileInfo;
-    const payload = atob(token[1]);
-    var ob = JSON.parse(payload);
-    const filepath = ob["file"];
-    const fileparts = filepath.split("/");
-    const filename = fileparts[fileparts.length-1];
-    const fileidx = filename.lastIndexOf('.');
-    var fileext = '';
-    var filetype = 'text';
-    if (fileidx > -1) {
-      fileext = filename.slice(fileidx);
-      switch (fileext) {
-        case '.c':
-        case '.h':
-        case '.inc':
-          filetype = 'c_cpp';
-          console.log("filetype:",filetype);
-      }
-    }
+  
+
+  processFileInfo( fileinfo : FileInfo) : FileInfo {
+    const url = fileinfo.lasturl;
+    const filepath = fileinfo.file;
+    this.logger.trace('FilesService-processFileInfo-start',fileinfo);
     if (this.filemap.hasOwnProperty(filepath)) {
-      fileinfo = this.filemap[filepath];
-      fileinfo.alreadyLoaded = true;
+      var cfileinfo : FileInfo = this.filemap[filepath];
+      if (fileinfo.saveActive) {
+        cfileinfo.save02_url(url);
+      } else {
+        cfileinfo.save05_error(cfileinfo.windowsId,"Vorhandene FileInfo");
+      }
+      cfileinfo.alreadyLoaded = true;
+      this.logger.trace('FilesService-processFileInfo-alreadyLoaded',cfileinfo);
+      return cfileinfo;
     } else {
-      fileinfo = new FileInfo();
-      fileinfo.filename = filename;
-      fileinfo.filepath = filepath;
-      fileinfo.filetype = filetype;
-      fileinfo.lasturl = url;
       fileinfo.saveActive = false;
       fileinfo.alreadyLoaded = false;
       this.filemap[filepath] = fileinfo;
     }
     var other = this;
     fileinfo.relateWindow = function(winid : string) {
+      other.logger.debug('FilesService-relaeWindow',winid);
       this.windowsId = winid;
-      other.filemap[filepath] = fileinfo;
+      other.filemap[filepath] = this;
     };
+    fileinfo.save02_url = function(url2) {
+      fileinfo.lasturl = url2;
+      other.logger.debug('FilesService-save02_url',fileinfo);
+      other.http.put(url2,fileinfo.content,{ responseType: 'text'}).subscribe((value:string) => {
+        fileinfo.oldContent = fileinfo.content;
+        fileinfo.save03_saved(filepath,function(err2,data){
+          other.logger.debug('FilesService-save03_saved',fileinfo,err2);
+          if (typeof(err2) !== "undefined") {
+            fileinfo.save05_error(fileinfo.windowsId,err2);
+            return;
+          }
+          if (fileinfo.temporary) {
+            fileinfo.save04_closing(fileinfo.windowsId);
+          } else {
+            fileinfo.save06_success(fileinfo.windowsId);
+          }
+        });
+      },(err:any) => {
+        other.logger.error('FilesService-save02_url-rrror',fileinfo,err);
+        fileinfo.save05_error(fileinfo.windowsId,err);
+      });
+    }
     fileinfo.load = function(cb) {
         other.http.get(url,{ responseType: 'text'}).subscribe((value:string) => {
-          console.log("File loaded: ",filepath);
+          other.logger.debug('FilesService-load',filepath);
           cb(undefined,value);
         },(err:any) => {
-          console.log("File load failed: ",filepath,err);
+          other.logger.debug('FilesService-load-failed',filepath,err);
           cb (err,undefined);
         });
       };
-    fileinfo.save = function(txt,cb) {
-        other.http.put(url,txt,{ responseType: 'text'}).subscribe((value:string) => {
-          console.log("File saved: ",filepath);
-          cb(undefined,value);
-        },(err:any) => {
-          console.log("File save failed: ",filepath,err);
-          cb (err,undefined);
-        });
-    };
     return fileinfo;
   }
 
-  constructor(private http: HttpClient) { }
+  constructor(
+    private logger:NGXLogger,private http: HttpClient) { }
 }